Fix SDK and JS testing bugs and improve web/Chrome compatibility
- AccountResponse.fromJson: swapped num_sponsoring/num_sponsored args
- contract_spec _inferType: Uint8List matched as List before Uint8List check
- TxRep parser: expected memo.return but encoder wrote memo.retHash
- KYC service: GET request builders did not pass custom headers
- ManageSellOffer/ManageBuyOffer: offerId .toInt().toString() lost precision on web
- removeTailZero: returned empty string for input "0" on web
- Extend XDR enum classes missing operator==/hashCode, causing potential switch
  statement failures on web for negative discriminant values
- Remove dead code (readLong, writeLong, u64BytesHelper, u64BigIntBytesHelper) and unused imports from xdr_data_io.dart.
- Fix web compatibility in tests: replace BigInt.from() with BigInt.parse() for literals exceeding JS safe integer range, accept platform-dependent double formatting, remove two native-only overflow tests.

+++ lib/src/manage_buy_offer_operation.dart
@@ -132,7 +132,7 @@ class ManageBuyOfferOperation extends Operation {
       Asset.fromXdr(op.buying),
       Util.fromXdrBigInt64Amount(op.amount.bigInt),
       removeTailZero((BigInt.from(n) / BigInt.from(d)).toString()),
-    ).setOfferId(op.offerID.uint64.toInt().toString());
+    ).setOfferId(op.offerID.uint64.toString());
   }
 }
