+++ lib/src/manage_sell_offer_operation.dart
@@ -131,7 +131,7 @@ class ManageSellOfferOperation extends Operation {
       Asset.fromXdr(op.buying),
       Util.fromXdrBigInt64Amount(op.amount.bigInt),
       removeTailZero((BigInt.from(n) / BigInt.from(d)).toString()),
-    ).setOfferId(op.offerID.uint64.toInt().toString());
+    ).setOfferId(op.offerID.uint64.toString());
   }
 }
